[cr132] `SimpleMenuModel` header moved to new path
    Chromium change:
    https://chromium.googlesource.com/chromium/src/+/f1147f74e2db4891bd99bd6a5d0feffe8b453ab2

    commit f1147f74e2db4891bd99bd6a5d0feffe8b453ab2
    Author: Joseph Park <josephjoopark@chromium.org>
    Date:   Fri Oct 18 17:39:17 2024 +0000

        Move SimpleMenuModel out of ui/base

        This is a precursor CL to allow SimpleMenuModel to include actions.h, in order to be able to create menu items from action items.
        Currently, simple_menu_model cannot include actions, as doing so creates a dependency cycle from ui/base -> ui/actions -> ui/base.

        Many of the changes here are simply renaming include paths from ui/base/models -> ui/menus/models. The only significant changes here that need to be reviewed are changes in BUILD.gn and DEPS files.
